+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.3.0 (in development)
+-----------------------
+- Formatting `InputArg::Stdin` or `OutputArg::Stdout` with `{:#}` will now
+  produce `"<stdin>"` or `"<stdout>"`, respectively
+
 v0.2.0 (2023-04-21)
 -------------------
 - Added "tokio" feature to enable async methods